python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
python-socketio==5.10.0
python-dotenv==1.0.0
slowapi==0.1.9
sqlalchemy==2.0.23
psycopg2-binary==2.9.9